pruebas

+++ PolizasApp/src/main/java/com/sura/polizas/bean/RespuestaBean.java
@@ -1,9 +1,11 @@
 package com.sura.polizas.bean;
 
+import java.math.BigDecimal;
+
 public class RespuestaBean {
 
 	private boolean financiado;
-	private Long valorPoliza;
+	private BigDecimal valorPoliza;
 
 	
 	public boolean isFinanciado() {
@@ -12,12 +14,14 @@ public class RespuestaBean {
 	public void setFinanciado(boolean financiado) {
 		this.financiado = financiado;
 	}
-	public Long getValorPoliza() {
+	public BigDecimal getValorPoliza() {
 		return valorPoliza;
 	}
-	public void setValorPoliza(Long valorPoliza) {
+	public void setValorPoliza(BigDecimal valorPoliza) {
 		this.valorPoliza = valorPoliza;
 	}
+	
+	
 	
 	
 }
